backend/requirements.txt
# ── Core inference ────────────────────────────────────────────────────────────
torch>=2.2.0,<3.0.0
transformers>=4.40.0
accelerate>=0.27.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# ── Quantization (optional — skip if no BnB support) ─────────────────────────
bitsandbytes>=0.43.0; platform_system != "Windows"

# ── RAG ───────────────────────────────────────────────────────────────────────
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
numpy>=1.24.0

# ── PDF extraction (optional) ─────────────────────────────────────────────────
pymupdf>=1.23.0

# ── API server ────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-multipart>=0.0.9

# ── Utilities ─────────────────────────────────────────────────────────────────
httpx>=0.27.0   # for testing
